+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `x-reduct-api` header to get quick version number in Major.Minor format, [PR-291](https://github.com/reductstore/reductstore/pull/291)
 - `GET /api/v1/:bucket/:entry/batch` endpoint to read a bunch of records, [PR-294](https://github.com/reductstore/reductstore/pull/294)
 
 ## [1.4.0] - 2023-06-09